refactor(core): Move base-element to elements/base directory.
