+++ core/src/elements/ons-col.js
@@ -15,7 +15,7 @@ limitations under the License.
 
 */
 
-import BaseElement from '../ons/base-element';
+import BaseElement from './base/base-element';
 
 /**
  * @element ons-col
