+++ core/src/elements/ons-icon.js
@@ -17,7 +17,7 @@ limitations under the License.
 
 import util from '../ons/util';
 import autoStyle from '../ons/autostyle';
-import BaseElement from '../ons/base-element';
+import BaseElement from './base/base-element';
 
 /**
  * @element ons-icon
