+++ core/src/elements/ons-if.js
@@ -17,7 +17,7 @@ limitations under the License.
 
 import orientation from '../ons/orientation';
 import platform from '../ons/platform';
-import BaseElement from '../ons/base-element';
+import BaseElement from './base/base-element';
 import contentReady from '../ons/content-ready';
 
 /**
