+++ core/src/elements/ons-lazy-repeat.js
@@ -11,7 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import BaseElement from '../ons/base-element';
+import BaseElement from './base/base-element';
 import util from '../ons/util';
 import {LazyRepeatDelegate, LazyRepeatProvider} from '../ons/internal/lazy-repeat';
