+++ core/src/elements/ons-list-header.js
@@ -18,7 +18,7 @@ limitations under the License.
 import util from '../ons/util';
 import autoStyle from '../ons/autostyle';
 import ModifierUtil from '../ons/internal/modifier-util';
-import BaseElement from '../ons/base-element';
+import BaseElement from './base/base-element';
 
 const defaultClassName = 'list-header';
 const scheme = {'': 'list-header--*'};
