+++ core/src/elements/ons-progress-circular.js
@@ -17,7 +17,7 @@ limitations under the License.
 
 import util from '../ons/util';
 import ModifierUtil from '../ons/internal/modifier-util';
-import BaseElement from '../ons/base-element';
+import BaseElement from './base/base-element';
 import contentReady from '../ons/content-ready';
 
 const scheme = {
