+++ core/src/elements/ons-pull-hook.js
@@ -17,7 +17,7 @@ limitations under the License.
 
 import util from '../ons/util';
 import platform from '../ons/platform';
-import BaseElement from '../ons/base-element';
+import BaseElement from './base/base-element';
 import GestureDetector from '../ons/gesture-detector';
 import animit from '../ons/animit';
